requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
asyncpg
aiomysql
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
alembic
aiosqlite
httpx[socks]
socksio
minio
pyyaml
python-multipart

# Sisyphus API Engine - 核心执行器
Sisyphus-api-engine==1.0.1
